jdbc update and delete refactoring

+++ webStudy01/src/kr/or/ddit/member/service/MemberServiceImpl.java
@@ -10,6 +10,7 @@ import kr.or.ddit.vo.MemberVO;
 
 public class MemberServiceImpl implements IMemberService {
 	private IMemberDAO dao = new MemberDAOImpl();
+	private IAuthenticateService authService = new AuthenticateServiceImpl();
 	
 	@Override
 	public MemberVO retrieveMember(String mem_id) {
@@ -38,32 +39,35 @@ public class MemberServiceImpl implements IMemberService {
 
 	@Override
 	public ServiceResult modifyMember(MemberVO member) {
-		MemberVO savedMember = retrieveMember(member.getMem_id());
-		String savedPass = savedMember.getMem_pass();
-		String inputPass = member.getMem_pass();
+		retrieveMember(member.getMem_id());	// 존재하지 않으면 예외발생용
+		ServiceResult result = 
+				authService.authenticate(new MemberVO(member.getMem_id(),member.getMem_pass()));
 		
-		ServiceResult result = null;
-		
-		if(savedPass.equals(inputPass)) {
+		if(ServiceResult.OK.equals(result)) {
 			int rowcnt = dao.updateMember(member);
 			if(rowcnt > 0) {
 				result = ServiceResult.OK;
 			}else {
 				result = ServiceResult.FAIL;
 			}
-		}else {
-			result = ServiceResult.INVALIDPASSWORD;
 		}
-		
 		return result;
 	}
 
 	@Override
 	public ServiceResult removeMember(MemberVO member) {
-		ServiceResult result = null;
+		retrieveMember(member.getMem_id());	// 존재하지 않으면 예외발생용
+		ServiceResult result = 
+				authService.authenticate(new MemberVO(member.getMem_id(),member.getMem_pass()));
 		
-		int rowcnt = dao.deleteMember(member.getMem_id());
-		
+		if(ServiceResult.OK.equals(result)) {
+			int rowcnt = dao.deleteMember(member.getMem_id());
+			if(rowcnt > 0) {
+				result = ServiceResult.OK;
+			}else {
+				result = ServiceResult.FAIL;
+			}
+		}
 		return result;
 	}
